* 'dev' of github.com:nickjhathaway/elucidatornf:
  check if sub regions exist
  update for subsegments analysis;
  increase default max_time;

+++ modules/local/get_sub_segments_from_PW_pop_res/main.nf
@@ -14,6 +14,8 @@ process GET_SUB_SEGMENTS_FROM_FASTA {
 
     script:
     def genome_fnp = "${genome_dir_fnp}/${primary_genome}.fasta"
+    def extra_args = task.ext.args ? task.ext.args : ''
+
     """
     elucidator createSharedSubSegmentsFromRefSeqs \
         --genome ${genome_fnp} \
@@ -23,12 +25,18 @@ process GET_SUB_SEGMENTS_FROM_FASTA {
         --dout subSegments \
         --overWriteDir \
         --correctionOccurenceCutOff ${correction_occurence_cut_off} \
-        --lowFreqCutOff ${low_freq_cut_off};
+        --lowFreqCutOff ${low_freq_cut_off} \
+        ${extra_args};
 
     # rename files so they have the target_id in them
-    cat "subSegments/subRegionInfo/0_ref_variable_expanded_genomic.bed" > "${region_id}_0_ref_variable_expanded_genomic.bed"
-    cat "subSegments/subRegionInfo/0_ref_sharedLocs_genomic.bed" > "${region_id}_0_ref_sharedLocs_genomic.bed"
-    cat "subSegments/subRegionInfo/0_ref_variable_genomic.bed" > "${region_id}_0_ref_variable_genomic.bed"
-
+    if [ -f subSegments/subRegionInfo/0_ref_variable_expanded_genomic.bed ]; then
+        cat "subSegments/subRegionInfo/0_ref_variable_expanded_genomic.bed" > "${region_id}_0_ref_variable_expanded_genomic.bed"
+        cat "subSegments/subRegionInfo/0_ref_sharedLocs_genomic.bed" > "${region_id}_0_ref_sharedLocs_genomic.bed"
+        cat "subSegments/subRegionInfo/0_ref_variable_genomic.bed" > "${region_id}_0_ref_variable_genomic.bed"
+    else
+        touch "${region_id}_0_ref_variable_expanded_genomic.bed"
+        touch "${region_id}_0_ref_sharedLocs_genomic.bed"
+        touch "${region_id}_0_ref_variable_genomic.bed"
+    fi
     """
 }
